Ads-Service/requirements.txt
bottle
googleads
google-cloud-firestore
